Add Printing of Program Advance Combos
When viewing Program Advances, each combination of chips used to
activate are now printed alongside the previously shown information. For
P.A.s that only possess one combination of chips, a unique version of
the line is printed. For P.A.s with multiple combinations, the P.A.
prints each one with commas where appropriate.